HsProof.Proof

Documentation

admitted :: Proof a

admit

qed :: Proof a

finish the proof

abort :: Proof a

abort the proof and return current goal

status :: Proof PropRef

show all proof content

exfalso :: ObjectId -> Proof ()

exfalso: F -> anything

apply' :: Appliable a Prop => a -> Proof ()

apply theorem/prop to the goal

apply :: ObjectId -> Proof ()

same as apply', but using ObjectId

applyToM :: Appliable a [Prop] => a -> [ObjectId] -> Proof ObjectId

apply to multiple props and create a new prop object

applyToM' :: Appliable a [Prop] => a -> [ObjectId] -> ObjectId -> Proof ()

applyM, but modify one instead of creating a new one

applyTo' :: Appliable a Prop => a -> ObjectId -> Proof ()

apply to a Prop

applyTo :: ObjectId -> ObjectId -> Proof ()

same as applyTo', but using ObjectId

intro :: Proof ObjectId

like intro in coq

imply :: Theorem'

applyTo acts differently, so we need a separate method

theorem' :: ObjectId -> Proof Theorem

lift a Prop to Theorem using ObjectId

theorem :: Prop -> Theorem

lifts a Prop to Theorem

(>$>) :: ObjectId -> ObjectId -> Proof Theorem

lifted (>=>) combining theorems: (a -> b) -> (b -> c) => (a -> c)